refactor(launcher): split orchestration and startup modules
